added clients dashboard and other functionality

+++ app/Http/Controllers/BlogController.php
@@ -8,11 +8,7 @@ use Auth;
 
 class BlogController extends Controller
 {
-    /**
-     * Display a listing of the resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
+  
     public function index()
     {
         if (Auth::check()) {
@@ -28,26 +24,18 @@ class BlogController extends Controller
         return view('blog', ['blogs' => $blogs]);
     }
 
-
-    /**
-     * Show the form for creating a new resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
     public function create()
     {
-        return view('Blogs.create');
+        if (Auth::check()) {
+            return view('Blogs.create');
+        }
+        return redirect()->route('login');
     }
 
-    /**
-     * Store a newly created resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @return \Illuminate\Http\Response
-     */
     public function store(Request $request)
     {
-        $file = $request->file('photo');
+        if (Auth::check()) {
+            $file = $request->file('photo');
             $name = str_replace(" ", "_", $request->title);
             $ext = $file->getClientOriginalExtension();
             if ($file->move("images/blogImages", $name.".".$ext)){
@@ -63,64 +51,46 @@ class BlogController extends Controller
                 }
             } 
             return "Error Uploading file";
+        } return redirect()->route('login');
     }
 
-    /**
-     * Display the specified resource.
-     *
-     * @param  \App\Blog  $blog
-     * @return \Illuminate\Http\Response
-     */
     public function show(Blog $blog)
     {
         $blog = Blog::find($blog->id);
         return view('showBlog', ['blog' => $blog]);
     }
 
-    /**
-     * Show the form for editing the specified resource.
-     *
-     * @param  \App\Blog  $blog
-     * @return \Illuminate\Http\Response
-     */
     public function edit(Blog $blog)
     {
-        $blog = Blog::find($blog->id);
-        return view('Blogs.edit', ['blog' => $blog]);
+        if (Auth::check()) {
+            $blog = Blog::find($blog->id);
+            return view('Blogs.edit', ['blog' => $blog]);
+        } return redirect()->route('login');
     }
 
-    /**
-     * Update the specified resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  \App\Blog  $blog
-     * @return \Illuminate\Http\Response
-     */
     public function update(Request $request)
     {
-        if (Blog::where('id', $request->id)->update([
+        if (Auth::check()) {
+            if (Blog::where('id', $request->id)->update([
             'title' => $request->title,
             'description' => $request->description
-        ])) {
+            ])) {
             return redirect()->route('blogs');
-        } 
-        return "Error";
+            } 
+            return "Error";
+        } return redirect()->route('login');
     }
 
-    /**
-     * Remove the specified resource from storage.
-     *
-     * @param  \App\Blog  $blog
-     * @return \Illuminate\Http\Response
-     */
     public function destroy($id)
     {
-        $blog = Blog::find($id);
-        if (unlink('images/blogImages/'.$blog->photo)){
-            if(Blog::where('id', $id)->delete()){
+        if (Auth::check()) {
+            $blog = Blog::find($id);
+            if (unlink('images/blogImages/'.$blog->photo)){
+                if(Blog::where('id', $id)->delete()){
                 return redirect()->route('blogs');
+                }
             }
-        }
-        return "Error";
+            return "Error";
+        } return redirect()->route('login');
     }
 }
